+++ cli/src/app.rs
@@ -254,6 +254,21 @@ pub fn add_subcommands(command: Command) -> Command {
                             .action(ArgAction::Append)
                             .help("Unique token names which identify the tokens"),
                     ),
+                )
+                .subcommand(
+                    Command::new("create-token").about("Create a new API token").args(&[
+                        Arg::new("token-name")
+                            .value_name("TOKEN_NAME")
+                            .action(ArgAction::Set)
+                            .required(true)
+                            .help("Unique name to identify the new token"),
+                        Arg::new("expiry")
+                            .value_name("DAYS")
+                            .short('e')
+                            .long("expiry")
+                            .action(ArgAction::Set)
+                            .help("Number of days the token will be valid"),
+                    ]),
                 ),
         )
         .subcommand(Command::new("ping").about("Ping the remote system to verify it is available"))