+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - `phylum auth list-tokens` subcommand to list API tokens
 - `phylum auth revoke-token` subcommand to revoke API tokens
+- `phylum auth create-token` subcommand to create API tokens
 
 ### Fixed
 - `Gemfile.lock` parsing with zero dependencies